+++ lib/nxt_support/util.rb
@@ -1,5 +1,5 @@
 require_relative 'util/enum'
-require_relative 'util/enum_collection'
+require_relative 'util/enum/value'
 require_relative 'util/hash_translator'
 require_relative 'util/crystalizer'
 require_relative 'util/birth_date'
