+++ lib/nxt_support/util/enum/value.rb
@@ -1,0 +1,7 @@
+module NxtSupport
+  class Enum::Value < String
+    def initialize(value)
+      super(value.to_s).freeze
+    end
+  end
+end
